services/comments/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.3
python-jose==3.3.0
pika==1.3.2
requests==2.32.4
pydantic==2.11.5
jwt==1.3.1